Moved implementation of menus in `Window` in separate module Moved public methods of `Window` to the top of implementation
